packages/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "taxlens-api"
version = "0.1.0"
description = "FastAPI backend for TaxLens"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-multipart>=0.0.9",
    "taxlens-engine",
    "python-jose[cryptography]>=3.3.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=7.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["app"]
